requirements.txt
accelerate==1.5.2
altair==5.5.0
attrs==25.3.0
bitsandbytes==0.45.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
diffusers==0.32.2
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.3
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.32.0
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
peft==0.15.1
pillow==11.1.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.0
streamlit-image-comparison==0.0.4
streamlit-image-select==0.6.0
streamlit-option-menu==0.4.0
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.3.13
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
xformers==0.0.29.post3
zipp==3.21.0
